feat: NoSessionException 추가, 각 Controller에 적용

+++ src/main/java/com/example/gistcompetitioncnserver/user/UserController.java
@@ -1,5 +1,6 @@
 package com.example.gistcompetitioncnserver.user;
 
+import com.example.gistcompetitioncnserver.exception.user.NoSessionException;
 import com.example.gistcompetitioncnserver.exception.user.UnAuthorizedUserException;
 import lombok.AllArgsConstructor;
 import org.springframework.http.ResponseEntity;
@@ -39,6 +40,9 @@ public class UserController {
     @GetMapping("/users")
     public ResponseEntity<List<User>> retrieveAllUsers() {
         SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (sessionUser == null) {
+            throw new NoSessionException();
+        }
         if (!sessionUser.isAdmin()) {
             throw new UnAuthorizedUserException();
         }
@@ -48,6 +52,9 @@ public class UserController {
     @GetMapping("/users/{userId}")
     public ResponseEntity<User> retrieveUser(@PathVariable Long userId) {
         SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (sessionUser == null) {
+            throw new NoSessionException();
+        }
         if (!sessionUser.isAdmin()) {
             throw new UnAuthorizedUserException();
         }
@@ -57,6 +64,9 @@ public class UserController {
     @GetMapping("/users/me")
     public ResponseEntity<User> retrieveUserOfMine() {
         SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (sessionUser == null) {
+            throw new NoSessionException();
+        }
         return ResponseEntity.ok().body(userService.findUserById(sessionUser.getId()));
     }
 
@@ -64,6 +74,9 @@ public class UserController {
     public ResponseEntity<Void> updateUserRole(@PathVariable Long userId,
                                                @Validated @RequestBody UpdateUserRoleRequest userRoleRequest) {
         SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (sessionUser == null) {
+            throw new NoSessionException();
+        }
         if (!sessionUser.isAdmin()) {
             throw new UnAuthorizedUserException();
         }
@@ -74,6 +87,9 @@ public class UserController {
     @PutMapping("/users/me/password")
     public ResponseEntity<Void> updatePasswordOfMine(@Validated @RequestBody UpdatePasswordRequest passwordRequest) {
         SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (sessionUser == null) {
+            throw new NoSessionException();
+        }
         userService.updatePassword(sessionUser.getId(), passwordRequest);
         return ResponseEntity.noContent().build();
     }
@@ -81,6 +97,9 @@ public class UserController {
     @DeleteMapping("/users/{userId}")
     public ResponseEntity<Void> deleteUser(@PathVariable Long userId) {
         SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (sessionUser == null) {
+            throw new NoSessionException();
+        }
         if (!sessionUser.isAdmin()) {
             throw new UnAuthorizedUserException();
         }
@@ -91,6 +110,9 @@ public class UserController {
     @DeleteMapping("/users/me")
     public ResponseEntity<Void> deleteUserOfMine(@Validated @RequestBody DeleteUserRequest deleteUserRequest) {
         SessionUser sessionUser = (SessionUser) httpSession.getAttribute("user");
+        if (sessionUser == null) {
+            throw new NoSessionException();
+        }
         userService.deleteUserOfMine(sessionUser.getId(), deleteUserRequest);
         return ResponseEntity.noContent().build();
     }
